more survey updating
