+++ python/lsst/sims/featureScheduler/surveys/__init__.py
@@ -1,2 +1,2 @@
 from .base_survey import *
-from 
+from .surveys import *
