+++ python/lsst/sims/featureScheduler/surveys/to_port.py
@@ -476,312 +476,6 @@ class Simple_greedy_survey_fields(BaseSurvey):
         return observations
 
 
-def rotx(theta, x, y, z):
-    """rotate the x,y,z points theta radians about x axis"""
-    sin_t = np.sin(theta)
-    cos_t = np.cos(theta)
-    xp = x
-    yp = y*cos_t+z*sin_t
-    zp = -y*sin_t+z*cos_t
-    return xp, yp, zp
-
-
-
-
-class Blob_survey(Greedy_survey_fields):
-    """Select observations in large, mostly contiuguous, blobs.
-    """
-    def __init__(self, basis_functions, basis_weights,
-                 extra_features=None, filtername='r', filter2='g',
-                 slew_approx=7.5, filter_change_approx=140.,
-                 read_approx=2., exptime=30., nexp=2,
-                 ideal_pair_time=22., min_pair_time=15.,
-                 search_radius =30., alt_max = 85., az_range=90.,
-                 smoothing_kernel=None, nside=default_nside,
-                 dither=True, seed=42, ignore_obs='ack',
-                 tag_fields=False, tag_map=None, tag_names=None,
-                 sun_alt_limit=-19., survey_note='blob',
-                 sitename='LSST'):
-        """
-        Parameters
-        ----------
-        filtername : str ('r')
-            The filter to observe in.
-        filter2 : str ('g')
-            The filter to pair with the first observation. If set to None, no pair
-            will be observed.
-        slew_approx : float (7.5)
-            The approximate slewtime between neerby fields (seconds). Used to calculate
-            how many observations can be taken in the desired time block.
-        filter_change_approx : float (140.)
-             The approximate time it takes to change filters (seconds).
-        ideal_pair_time : float (22.)
-            The ideal time gap wanted between observations to the same pointing (minutes)
-        min_pair_time : float (15.)
-            The minimum acceptable pair time (minutes)
-        search_radius : float (30.)
-            The radius around the reward peak to look for additional potential pointings (degrees)
-        alt_max : float (85.)
-            The maximum altitude to include (degrees).
-        az_range : float (90.)
-            The range of azimuths to consider around the peak reward value (degrees).
-        sitename : str ('LSST')
-            The name of the site to lookup latitude and longitude.
-        """
-
-        if nside is None:
-            nside = set_default_nside()
-
-        if extra_features is None:
-            extra_features = {}
-            extra_features['night'] = features.Current_night()
-            extra_features['mounted_filters'] = features.Mounted_filters()
-            extra_features['mjd'] = features.Current_mjd()
-            extra_features['night_boundaries'] = features.CurrentNightBoundaries()
-            extra_features['sun_moon_alt'] = features.Sun_moon_alts()
-            extra_features['lmst'] = features.Current_lmst()  # Pretty sure in hours
-            extra_features['current_filter'] = features.Current_filter()
-            extra_features['altaz'] = features.AltAzFeature()
-
-        super(Blob_survey, self).__init__(basis_functions=basis_functions,
-                                          basis_weights=basis_weights,
-                                          extra_features=extra_features,
-                                          filtername=filtername,
-                                          block_size=0, smoothing_kernel=smoothing_kernel,
-                                          dither=dither, seed=seed, ignore_obs=ignore_obs,
-                                          tag_fields=tag_fields, tag_map=tag_map,
-                                          tag_names=tag_names,
-                                          nside=nside)
-        self.nexp = nexp
-        self.exptime = exptime
-        self.slew_approx = slew_approx
-        self.read_approx = read_approx
-        self.hpids = np.arange(hp.nside2npix(self.nside))
-        # If we are taking pairs in same filter, no need to add filter change time.
-        if filtername == filter2:
-            filter_change_approx = 0
-        # Compute the minimum time needed to observe a blob (or observe, then repeat.)
-        if filter2 is not None:
-            self.time_needed = (min_pair_time*60.*2. + exptime + read_approx + filter_change_approx)/24./3600.  # Days
-        else:
-            self.time_needed = (min_pair_time*60. + exptime + read_approx)/24./3600.  # Days
-        self.filter_set = set(filtername)
-        if filter2 is None:
-            self.filter2_set = self.filter_set
-        else:
-            self.filter2_set = set(filter2)
-        self.sun_alt_limit = np.radians(sun_alt_limit)
-
-        self.ra, self.dec = _hpid2RaDec(self.nside, self.hpids)
-        # Look up latitude and longitude for alt,az conversions later
-        # XXX: TODO: lat and lon should be in the Observatory feature. But that feature
-        # needs documentation on what's in it!
-        site = Site(name=sitename)
-        self.lat = site.latitude_rad
-        self.lon = site.longitude_rad
-        self.survey_note = survey_note
-        self.counter = 1  # start at 1, because 0 is default in empty observation
-        self.filter2 = filter2
-        self.search_radius = np.radians(search_radius)
-        self.az_range = np.radians(az_range)
-        self.alt_max = np.radians(alt_max)
-        self.min_pair_time = min_pair_time
-        self.ideal_pair_time = ideal_pair_time
-
-    def _set_block_size(self):
-        """
-        Update the block size if it's getting near the end of the night.
-        """
-
-        available_time = self.extra_features['night_boundaries'].feature['next_twilight_start'] -\
-                         self.extra_features['mjd'].feature
-        available_time *= 24.*60.  # to minutes
-
-        n_ideal_blocks = available_time / self.ideal_pair_time
-        if n_ideal_blocks >= 3:
-            self.nvisit_block = int(np.floor(self.ideal_pair_time*60. / (self.slew_approx + self.exptime +
-                                                                         self.read_approx*(self.nexp - 1))))
-        else:
-            # Now we can stretch or contract the block size to allocate the remainder time until twilight starts
-            # We can take the remaining time and try to do 1,2, or 3 blocks.
-            possible_times = available_time / np.arange(1, 4)
-            diff = np.abs(self.ideal_pair_time-possible_times)
-            best_block_time = np.max(possible_times[np.where(diff == np.min(diff))])
-            self.nvisit_block = int(np.floor(best_block_time*60. / (self.slew_approx + self.exptime +
-                                                                    self.read_approx*(self.nexp - 1))))
-
-    def _check_feasability(self):
-        # Check if filters are loaded
-        filters_mounted = self.filter_set.issubset(set(self.extra_features['mounted_filters'].feature))
-        if self.filter2 is not None:
-            second_fitler_mounted = self.filter2_set.issubset(set(self.extra_features['mounted_filters'].feature))
-            filters_mounted = filters_mounted & second_fitler_mounted
-
-        available_time = self.extra_features['night_boundaries'].feature['next_twilight_start'] - self.extra_features['mjd'].feature
-        if not filters_mounted:
-            return False
-        # Check we are not in twilight
-        elif self.extra_features['sun_moon_alt'].feature['sunAlt'] > self.sun_alt_limit:
-            return False
-        # We have enough time before twilight starts
-        elif available_time < self.time_needed:
-            return False
-        else:
-            return True
-
-    def calc_reward_function(self):
-        """
-        
-        """
-        # Set the number of observations we are going to try and take
-        self._set_block_size()
-        #  Computing reward like usual with basis functions and weights
-        if self._check_feasability():
-            self.reward = 0
-            indx = np.arange(hp.nside2npix(self.nside))
-            # keep track of masked pixels
-            mask = np.zeros(indx.size, dtype=bool)
-            for bf, weight in zip(self.basis_functions, self.basis_weights):
-                basis_value = bf(indx=indx)
-                mask[np.where(basis_value == hp.UNSEEN)] = True
-                if hasattr(basis_value, 'mask'):
-                    mask[np.where(basis_value.mask == True)] = True
-                self.reward += basis_value*weight
-                # might be faster to pull this out into the feasabiliity check?
-                if hasattr(self.reward, 'mask'):
-                    indx = np.where(self.reward.mask == False)[0]
-            self.reward[mask] = hp.UNSEEN
-            self.reward.mask = mask
-            self.reward.fill_value = hp.UNSEEN
-
-            if self.smoothing_kernel is not None:
-                self.smooth_reward()
-
-            # Apply max altitude cut
-            too_high = np.where(self.extra_features['altaz'].feature['alt'] > self.alt_max)
-            self.reward[too_high] = hp.UNSEEN
-
-            # Select healpixels within some radius of the max
-            # This is probably faster with a kd-tree.
-            peak_reward = np.min(np.where(self.reward == np.max(self.reward))[0])
-            # Apply radius selection
-            dists = haversine(self.ra[peak_reward], self.dec[peak_reward], self.ra, self.dec)
-            out_hp = np.where(dists > self.search_radius)
-            self.reward[out_hp] = hp.UNSEEN
-
-            # Apply az cut
-            az_centered = self.extra_features['altaz'].feature['az'] - self.extra_features['altaz'].feature['az'][peak_reward]
-            az_centered[np.where(az_centered < 0)] += 2.*np.pi
-
-            az_out = np.where((az_centered > self.az_range/2.) & (az_centered < 2.*np.pi-self.az_range/2.))
-            self.reward[az_out] = hp.UNSEEN
-            potential_hp = np.where(self.reward.filled() != hp.UNSEEN)
-            # Find the max reward for each potential pointing
-            ufields, reward_by_field = int_binned_stat(self.hp2fields[potential_hp],
-                                                       self.reward[potential_hp].filled(),
-                                                       statistic=max_reject)
-            order = np.argsort(reward_by_field)
-            ufields = ufields[order][::-1][0:self.nvisit_block]
-            self.best_fields = ufields
-        else:
-            self.reward = -np.inf
-        self.reward_checked = True
-        return self.reward
-
-    def __call__(self):
-        """
-        Find a good block of observations.
-        """
-        if not self.reward_checked:
-            # This should set self.best_fields
-            self.reward = self.calc_reward_function()
-
-        # Let's find the alt, az coords of the points (right now, hopefully doesn't change much in time block)
-        pointing_alt, pointing_az = _approx_RaDec2AltAz(self.fields['RA'][self.best_fields],
-                                                        self.fields['dec'][self.best_fields],
-                                                        self.lat, self.lon,
-                                                        self.extra_features['mjd'].feature,
-                                                        lmst=self.extra_features['lmst'].feature)
-        # Let's find a good spot to project the points to a plane
-        mid_alt = (np.max(pointing_alt) - np.min(pointing_alt))/2.
-
-        # Code snippet from MAF for computing mean of angle accounting for wrap around
-        # XXX-TODO: Maybe move this to sims_utils as a generally useful snippet.
-        x = np.cos(pointing_az)
-        y = np.sin(pointing_az)
-        meanx = np.mean(x)
-        meany = np.mean(y)
-        angle = np.arctan2(meany, meanx)
-        radius = np.sqrt(meanx**2 + meany**2)
-        mid_az = angle % (2.*np.pi)
-        if radius < 0.1:
-            mid_az = np.pi
-
-        # Project the alt,az coordinates to a plane. Could consider scaling things to represent
-        # time between points rather than angular distance.
-        pointing_x, pointing_y = gnomonic_project_toxy(pointing_az, pointing_alt, mid_az, mid_alt)
-        # Now I have a bunch of x,y pointings. Drop into TSP solver to get an effiencent route
-        towns = np.vstack((pointing_x, pointing_y)).T
-        # Leaving optimize=False for speed. The optimization step doesn't usually improve much.
-        better_order = tsp_convex(towns, optimize=False)
-        # XXX-TODO: Could try to roll better_order to start at the nearest/fastest slew from current position.
-        observations = []
-        counter2 = 0
-        for indx in better_order:
-            field = self.best_fields[indx]
-            if self.tag_fields:
-                tag = np.unique(self.tag_map[np.where(self.hp2fields == field)])[0]
-            else:
-                tag = 1
-            if tag == 0:
-                continue
-            obs = empty_observation()
-            obs['RA'] = self.fields['RA'][field]
-            obs['dec'] = self.fields['dec'][field]
-            obs['rotSkyPos'] = 0.
-            obs['filter'] = self.filtername
-            obs['nexp'] = self.nexp
-            obs['exptime'] = self.exptime
-            obs['field_id'] = -1
-            if self.tag_fields:
-                obs['survey_id'] = np.unique(self.tag_map[np.where(self.hp2fields == field)])[0]
-            else:
-                obs['survey_id'] = 1
-            obs['note'] = '%s' % (self.survey_note)
-            obs['block_id'] = self.counter
-            observations.append(obs)
-            counter2 += 1
-
-        # If we only want one filter block
-        if self.filter2 is None:
-            result = observations
-        else:
-            # Double the list to get a pair.
-            observations_paired = []
-            for observation in observations:
-                obs = copy.copy(observation)
-                obs['filter'] = self.filter2
-                observations_paired.append(obs)
-
-            # Check loaded filter here to decide which goes first
-            if self.extra_features['current_filter'].feature == self.filter2:
-                result = observations_paired + observations
-            else:
-                result = observations + observations_paired
-
-            # Let's tag which one is supposed to be first/second in the pair:
-            for i in range(0, int(np.size(result)/2), 1):
-                result[i]['note'] = '%s, a' % (self.survey_note)
-            for i in range(int(np.size(result)/2), np.size(result), 1):
-                result[i]['note'] = '%s, b' % (self.survey_note)
-
-        # XXX--note, we could run this like the DD surveys and keep the queue locally,
-        # then we can control how to recover if interupted, cut a sequence short if needed, etc.
-        # But that is a lot of code for what should be a minor improvement.
-
-        # Keep track of which block we're on. Nice for debugging.
-        self.counter += 1
-        return result
 
 
 class Block_survey(Greedy_survey_fields):
@@ -1215,331 +909,6 @@ def wrapHA(HA):
     return HA % 24.
 
 
-class Deep_drilling_survey(BaseSurvey):
-    """A survey class for running deep drilling fields
-    """
-    # XXX--maybe should switch back to taking basis functions and weights to
-    # make it easier to put in masks for moon and limits for seeing?
-    def __init__(self, RA, dec, extra_features=None, sequence='rgizy',
-                 nvis=[20, 10, 20, 26, 20],
-                 exptime=30.,
-                 nexp=2, ignore_obs='dummy', survey_name='DD', fraction_limit=0.01,
-                 ha_limits=([0., 1.5], [21.0, 24.]), reward_value=101., moon_up=True, readtime=2.,
-                 avoid_same_day=False,
-                 day_space=2., max_clouds=0.7, moon_distance=30., filter_goals=None, nside=default_nside):
-        """
-        Parameters
-        ----------
-        RA : float
-            The RA of the field (degrees)
-        dec : float
-            The dec of the field to observe (degrees)
-        extra_features : list of feature objects (None)
-            The features to track, will construct automatically if None.
-        sequence : list of observation objects or str (rgizy)
-            The sequence of observations to take. Can be a string of list of obs objects.
-        nvis : list of ints
-            The number of visits in each filter. Should be same length as sequence.
-        survey_name : str (DD)
-            The name to give this survey so it can be tracked
-        fraction_limit : float (0.01)
-            Do not request observations if the fraction of observations from this
-            survey exceeds the frac_limit.
-        ha_limits : list of floats ([-1.5, 1.])
-            The range of acceptable hour angles to start a sequence (hours)
-        reward_value : float (101.)
-            The reward value to report if it is able to start (unitless).
-        moon_up : bool (True)
-            Require the moon to be up (True) or down (False) or either (None).
-        readtime : float (2.)
-            Readout time for computing approximate time of observing the sequence. (seconds)
-        day_space : float (2.)
-            Demand this much spacing between trying to launch a sequence (days)
-        max_clouds : float (0.7)
-            Maximum allowed cloud value for an observation.
-        """
-        # No basis functions for this survey
-        basis_functions = []
-        basis_weights = []
-        self.ra = np.radians(RA)
-        self.ra_hours = RA/360.*24.
-        self.dec = np.radians(dec)
-        self.ignore_obs = ignore_obs
-        self.survey_name = survey_name
-        self.HA_limits = np.array(ha_limits)
-        self.reward_value = reward_value
-        self.moon_up = moon_up
-        self.fraction_limit = fraction_limit
-        self.day_space = day_space
-        self.survey_id = 5
-        self.nside = nside
-        self.filter_list = []
-        self.max_clouds = max_clouds
-        self.moon_distance = np.radians(moon_distance)
-        self.sequence = True  # Specifies the survey gives sequence of observations
-        self.avoid_same_day = avoid_same_day
-        self.filter_goals = filter_goals
-
-        if extra_features is None:
-            self.extra_features = {}
-            # Current filter
-            self.extra_features['current_filter'] = features.Current_filter()
-            # Available filters
-            self.extra_features['mounted_filters'] = features.Mounted_filters()
-            # Observatory information
-            self.extra_features['observatory'] = features.Observatory({'readtime': readtime,
-                                                                       'filter_change_time': 120.}
-                                                                      )  # FIXME:
-            self.extra_features['night'] = features.Current_night()
-            # The total number of observations
-            self.extra_features['N_obs'] = features.N_obs_count()
-            # The number of observations for this survey
-            self.extra_features['N_obs_self'] = features.N_obs_survey(note=survey_name)
-            # The current LMST. Pretty sure in hours
-            self.extra_features['lmst'] = features.Current_lmst()
-            # Moon altitude
-            self.extra_features['sun_moon_alt'] = features.Sun_moon_alts()
-            # Moon altitude
-            self.extra_features['moon'] = features.Moon()
-
-            # Time to next moon rise
-
-            # Time to twilight
-
-            # last time this survey was observed (in case we want to force a cadence)
-            self.extra_features['last_obs_self'] = features.Last_observation(survey_name=self.survey_name)
-            # last time a sequence observation
-            self.extra_features['last_seq_obs'] = features.LastSequence_observation(sequence_ids=[self.survey_id])
-
-            # Current MJD
-            self.extra_features['mjd'] = features.Current_mjd()
-            # Observable time. This includes altitude and night limits
-            # Fixme: add proper altitude limit from ha limits
-            self.extra_features['night_boundaries'] = features.CurrentNightBoundaries()
-            # Proposal information
-            self.extra_features['proposals'] = features.SurveyProposals(ids=(self.survey_id,),
-                                                                        names=(self.survey_name,))
-            # Cloud cover information
-            self.extra_features['bulk_cloud'] = features.BulkCloudCover()
-        else:
-            self.extra_features = extra_features
-
-        super(Deep_drilling_survey, self).__init__(basis_functions=basis_functions,
-                                                   basis_weights=basis_weights,
-                                                   extra_features=self.extra_features,
-                                                   nside=nside)
-
-        if type(sequence) == str:
-            opsim_fields = read_fields()
-            self.pointing2hpindx = hp_in_lsst_fov(nside=self.nside)
-            hp2fields = np.zeros(hp.nside2npix(self.nside), dtype=np.int)
-            for i in range(len(opsim_fields['RA'])):
-                hpindx = self.pointing2hpindx(opsim_fields['RA'][i], opsim_fields['dec'][i])
-                hp2fields[hpindx] = i+1
-            hpid = _raDec2Hpid(self.nside, self.ra, self.dec)
-
-            fields = read_fields()
-            field = fields[hp2fields[hpid]]
-            field['tag'] = self.survey_id
-            self.fields = [field]
-
-            self.sequence = []
-            self.sequence_dict = dict()
-            filter_list = []
-            for num, filtername in zip(nvis, sequence):
-                filter_list.append(filtername)
-                if filtername not in self.sequence_dict:
-                    self.sequence_dict[filtername] = []
-
-                for j in range(num):
-                    obs = empty_observation()
-                    obs['filter'] = filtername
-                    obs['exptime'] = exptime
-                    obs['RA'] = self.ra
-                    obs['dec'] = self.dec
-                    obs['nexp'] = nexp
-                    obs['note'] = survey_name
-                    obs['field_id'] = hp2fields[hpid]
-                    obs['survey_id'] = self.survey_id
-
-                    # self.sequence.append(obs)
-                    self.sequence_dict[filtername].append(obs)
-            self.filter_list = np.unique(np.array(filter_list))
-        else:
-            self.sequence_dict = None
-            self.sequence = sequence
-
-        # add extra features to map filter goals
-        for filtername in self.filter_list:
-            self.extra_features['N_obs_%s' % filtername] = features.N_obs_count(filtername=filtername)
-
-        self.approx_time = np.sum([(o['exptime']+readtime)*o['nexp'] for o in obs])
-
-        # Construct list of all the filters that need to be loaded to execute sequence
-        self.filter_set = set(self.filter_list)
-
-    def _check_feasability(self):
-        # Check that all filters are available
-        result = self.filter_set.issubset(set(self.extra_features['mounted_filters'].feature))
-        if not result:
-            return False
-
-        if (self.avoid_same_day and
-                (self.extra_features['last_seq_obs'].feature['night'] == self.extra_features['night'].feature)):
-            return False
-        # Check if the LMST is in range
-        HA = self.extra_features['lmst'].feature - self.ra_hours
-        HA = wrapHA(HA)
-
-        result = False
-        for limit in self.HA_limits:
-            lres = limit[0] <= HA < limit[1]
-            result = result or lres
-
-        if not result:
-            return False
-        # Check moon alt
-        if self.moon_up is not None:
-            if self.moon_up:
-                if self.extra_features['sun_moon_alt'].feature['moonAlt'] < 0.:
-                    return False
-            else:
-                if self.extra_features['sun_moon_alt'].feature['moonAlt'] > 0.:
-                    return False
-
-        # Make sure twilight hasn't started
-        if self.extra_features['sun_moon_alt'].feature['sunAlt'] > np.radians(-18.):
-            return False
-
-        # Check that it's been long enough since last sequence
-        if self.extra_features['mjd'].feature - self.extra_features['last_obs_self'].feature['mjd'] < self.day_space:
-            return False
-
-        # TODO: Check if the moon will come up. Compare next moonrise time to self.apporox time
-
-        # TODO: Check if twilight starts soon
-
-        # TODO: Make sure it is possible to complete the sequence of observations. Hit any limit?
-
-        # Check if there's still enough time to complete the observation
-        time_left = (self.extra_features['night_boundaries'].feature['next_twilight_start'] -
-                     self.extra_features['mjd'].feature) * 24.*60.*60.  # convert to seconds
-
-        seq_time = 42.  # Make sure there is enough time for an extra visit after the DD sequence
-        current_filter = self.extra_features['current_filter'].feature
-        for obs in self.sequence:
-            for o in obs:
-                if current_filter != o['filter']:
-                    seq_time += self.extra_features['observatory'].feature['filter_change_time']
-                    current_filter = o['filter']
-                seq_time += o['exptime']+self.extra_features['observatory'].feature['readtime']*o['nexp']
-
-        # log.debug('Time left: %.2f | Approx. time: %.2f' % (time_left, seq_time))
-        if time_left < seq_time:
-            return False
-
-        if self.extra_features['N_obs'].feature == 0:
-            return True
-
-        # Check if we are over-observed relative to the fraction of time alloted.
-        if self.extra_features['N_obs_self'].feature/float(self.extra_features['N_obs'].feature) > self.fraction_limit:
-            return False
-
-        # Check clouds
-        if self.extra_features['bulk_cloud'].feature > self.max_clouds:
-            return False
-
-        # If we made it this far, good to go
-        return result
-
-    def check_feasibility(self, observation):
-        '''
-        This method enables external calls to check if a given observations that belongs to this survey is
-        feasible or not. This is called once a sequence has started to make sure it can continue.
-
-        :return:
-        '''
-
-        # Check moon distance
-        if self.moon_up is not None:
-            moon_separation = _angularSeparation(self.extra_features['moon'].feature['moonRA'],
-                                                 self.extra_features['moon'].feature['moonDec'],
-                                                 observation['RA'],
-                                                 observation['dec'])
-            if moon_separation < self.moon_distance:
-                return False
-
-        # Check clouds
-        if self.extra_features['bulk_cloud'].feature > self.max_clouds:
-            return False
-
-        # If we made it this far, good to go
-        return True
-
-    def get_sequence(self):
-        '''
-        Build and return sequence of observations
-        :return:
-        '''
-
-        if self.sequence_dict is None:
-            return self.sequence
-        elif len(self.filter_list) == 1:
-            return self.sequence_dict[self.filter_list[0]]
-        elif self.filter_goals is None:
-            self.sequence = []
-            for filtername in self.filter_list:
-                for observation in self.sequence_dict[filtername]:
-                    self.sequence.append(observation)
-            return self.sequence
-
-        # If arrived here, then need to construct sequence. Will but the current filter first and the one that
-        # requires more observations last
-        filter_need = np.zeros(len(self.filter_list))
-        filter_goal = np.array([self.filter_goals[fname] for fname in self.filter_list])
-        filter_goal = (1.-filter_goal)/(1.+filter_goal)
-
-        if self.extra_features['N_obs'].feature > 0:
-            for i, filtername in enumerate(self.filter_list):
-                filter_need[i] = ((self.extra_features['N_obs'].feature -
-                                   self.extra_features['N_obs_%s' % filtername].feature) /
-                                  (self.extra_features['N_obs'].feature +
-                                   self.extra_features['N_obs_%s' % filtername].feature)) / filter_goal[i]
-        else:
-            filter_need = 1./filter_goal
-
-        filter_order = np.array(self.filter_list[np.argsort(filter_need)[::-1]])
-        current_filter_index = np.where(filter_order == self.extra_features['current_filter'].feature)[0]
-        if current_filter_index != 0 and current_filter_index != len(self.filter_list)-1:
-            first_filter = filter_order[0]
-            filter_order[0] = self.extra_features['current_filter'].feature
-            filter_order[current_filter_index] = first_filter
-        elif current_filter_index == len(self.filter_list)-1:
-            filter_order = np.append(filter_order[-1], filter_order[:-1])
-
-        log.debug('DeepDrilling[filter_order]: %s was %s[need: %s] ' % (filter_order,
-                                                                        self.filter_list,
-                                                                        filter_need))
-        self.sequence = []
-        for filtername in filter_order:
-            for observation in self.sequence_dict[filtername]:
-                self.sequence.append(observation)
-        return self.sequence
-
-
-    def calc_reward_function(self):
-        result = -np.inf
-        if self._check_feasability():
-            result = self.reward_value
-        return result
-
-    def __call__(self):
-        result = []
-        if self._check_feasability():
-            result = copy.deepcopy(self.get_sequence())
-            # Note, could check here what the current filter is and re-order the result
-        return result
-
 
 class Pairs_survey_scripted(Scripted_survey):
     """Check if incoming observations will need a pair in 30 minutes. If so, add to the queue
@@ -1833,65 +1202,6 @@ class Pairs_different_filters_scripted(Pairs_survey_scripted):
         return result
 
 
-def generate_dd_surveys(nside=default_nside):
-    """Utility to return a list of standard deep drilling field surveys.
-
-    XXX-Someone double check that I got the coordinates right!
-
-    XXX--I suspect that scheduling the DD fields all simultaneously ahead of
-    time would be a better strategy, so this should become obsolete.
-    """
-
-    surveys = []
-    # ELAIS S1
-    surveys.append(Deep_drilling_survey(9.45, -44., sequence='rgizy',
-                                        nvis=[20, 10, 20, 26, 20],
-                                        survey_name='DD:ELAISS1', reward_value=100, moon_up=None,
-                                        fraction_limit=0.0185, ha_limits=([0., 1.18], [21.82, 24.]),
-                                        nside=nside))
-    surveys.append(Deep_drilling_survey(9.45, -44., sequence='u',
-                                        nvis=[7],
-                                        survey_name='DD:u,ELAISS1', reward_value=100, moon_up=False,
-                                        fraction_limit=0.0015, ha_limits=([0., 1.18], [21.82, 24.]),
-                                        nside=nside))
-
-    # XMM-LSS
-    surveys.append(Deep_drilling_survey(35.708333, -4-45/60., sequence='rgizy',
-                                        nvis=[20, 10, 20, 26, 20],
-                                        survey_name='DD:XMM-LSS', reward_value=100, moon_up=None,
-                                        fraction_limit=0.0185, ha_limits=([0., 1.3], [21.7, 24.]),
-                                        nside=nside))
-    surveys.append(Deep_drilling_survey(35.708333, -4-45/60., sequence='u',
-                                        nvis=[7],
-                                        survey_name='DD:u,XMM-LSS', reward_value=100, moon_up=False,
-                                        fraction_limit=0.0015, ha_limits=([0., 1.3], [21.7, 24.]),
-                                        nside=nside))
-
-    # Extended Chandra Deep Field South
-    # XXX--Note, this one can pass near zenith. Should go back and add better planning on this.
-    surveys.append(Deep_drilling_survey(53.125, -28.-6/60., sequence='rgizy',
-                                        nvis=[20, 10, 20, 26, 20],
-                                        survey_name='DD:ECDFS', reward_value=100, moon_up=None,
-                                        fraction_limit=0.0185, ha_limits=[[0.5, 3.0], [20., 22.5]],
-                                        nside=nside))
-    surveys.append(Deep_drilling_survey(53.125, -28.-6/60., sequence='u',
-                                        nvis=[7],
-                                        survey_name='DD:u,ECDFS', reward_value=100, moon_up=False,
-                                        fraction_limit=0.0015, ha_limits=[[0.5, 3.0], [20., 22.5]],
-                                        nside=nside))
-    # COSMOS
-    surveys.append(Deep_drilling_survey(150.1, 2.+10./60.+55/3600., sequence='rgizy',
-                                        nvis=[20, 10, 20, 26, 20],
-                                        survey_name='DD:COSMOS', reward_value=100, moon_up=None,
-                                        fraction_limit=0.0185, ha_limits=([0., 1.5], [21.5, 24.]),
-                                        nside=nside))
-    surveys.append(Deep_drilling_survey(150.1, 2.+10./60.+55/3600., sequence='u',
-                                        nvis=[7], ha_limits=([0., 1.5], [21.5, 24.]),
-                                        survey_name='DD:u,COSMOS', reward_value=100, moon_up=False,
-                                        fraction_limit=0.0015,
-                                        nside=nside))
-
-    return surveys
 
 
 class Vary_expt_survey(Greedy_survey_fields):
